Create battleship.py
Battleship is a  game in which the player has to guess the coordinates of the battleship with limited chances. If your guess is right, Battleship is said to be sunken.. If you exceed the no. of chances, the game will be over.
This program is good for practice.. If you're a beginner in  python .. then better code this game ..